Other: Extract `createButtonForDropdown()` to own file.
